Segmentation fault fixed. New problem: infinity loop

+++ pomcp/src/history.h
@@ -1,3 +1,4 @@
+
 #ifndef HISTORY_H
 #define HISTORY_H
 
@@ -25,7 +26,7 @@ public:
     {
         if (history.History.size() != History.size())
             return false;
-        for (int i = 0; i < History.size(); ++i)
+        for (size_t i = 0; i < History.size(); ++i)
             if (history.History[i].Action != History[i].Action
              || history.History[i].Observation != History[i].Observation)
                 return false;
@@ -83,7 +84,7 @@ public:
 
     void Display(std::ostream& ostr) const
     {
-        for (int t = 0; t < History.size(); ++t)
+        for (size_t t = 0; t < History.size(); ++t)
         {
             ostr << "a=" << History[t].Action <<  " ";
             if (History[t].Observation >= 0)
